+++ src/game/maze/list.cpp
@@ -117,7 +117,7 @@ bool CellList::hasSameIndex()
 {
     int index = list[0].index;
 
-    for (int i = 1; i < list.size(); i++) {
+    for (size_t i = 1; i < list.size(); i++) {
         if (list[i].index != index) {
             return false;
         }
